interface: fix circular dependencies
