+++ pkg/interface/src/views/components/MentionText.tsx
@@ -2,8 +2,9 @@ import { Text } from '@tlon/indigo-react';
 import { Contact, Content, Group } from '@urbit/api';
 import React from 'react';
 import { referenceToPermalink } from '~/logic/lib/permalinks';
-import { cite, deSig, useShowNickname } from '~/logic/lib/util';
+import { cite, deSig } from '~/logic/lib/util';
 import { useContact } from '~/logic/state/contact';
+import { useShowNickname } from '~/logic/state/settings';
 import { PropFunc } from '~/types';
 import ProfileOverlay from '~/views/components/ProfileOverlay';
 import RichText from '~/views/components/RichText';
